# openapi.json
{"openapi":"3.0.2","info":{"title":"KidEdVisor","description":"`Kids education searching`","version":"0.0.2"},"paths":{"/api/v1/auth/registration":{"post":{"tags":["AUTH"],"summary":"Registration a new user","description":"User data is temporarily saved. After registration, the user receive a confirmation link to his email. The data will be deleted if the user doesn't confirm the registration.","operationId":"registration_api_v1_auth_registration_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTempUserScheme"}}},"required":true},"responses":{"202":{"description":"Successful Response returns only status code 202","content":{"application/json":{"schema":{}}}},"422":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}},"400":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}}}}},"/api/v1/auth/confirm/{uuid}":{"get":{"tags":["AUTH"],"summary":"User registration confirmation","description":"The user will be removed from temporary storage and saved as a real user.","operationId":"confirm_registration_api_v1_auth_confirm__uuid__get","parameters":[{"description":"Http link suffix for user registration","required":true,"schema":{"title":"Uuid","maxLength":32,"minLength":32,"type":"string","description":"Http link suffix for user registration"},"example":"993418d6434848bc927a13c48ca111f4","name":"uuid","in":"path"}],"responses":{"201":{"description":"Successful Response returns only status code 201","content":{"application/json":{"schema":{}}}},"422":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}},"400":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}}}}},"/api/v1/auth/token":{"post":{"tags":["AUTH"],"summary":"Get authorization token","description":"User submits form with email password, user type and gets `JWT` token","operationId":"login_api_v1_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_v1_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenScheme"}}}},"422":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}},"400":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}}}}},"/api/v1/auth/change_password":{"post":{"tags":["AUTH"],"summary":"Change password","description":"The authorized user submits a form with a new password.","operationId":"request_change_password_api_v1_auth_change_password_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_request_change_password_api_v1_auth_change_password_post"}}},"required":true},"responses":{"200":{"description":"Successful Response returns only status code 200","content":{"application/json":{"schema":{}}}},"422":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}},"400":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/auth/forget_password":{"post":{"tags":["AUTH"],"summary":"Get a link to create a new password","description":"The user specifies registration email address and receives a link to this email address. Then he follows this link and receives a new password.","operationId":"forget_password_api_v1_auth_forget_password_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_forget_password_api_v1_auth_forget_password_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}},"400":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}}}}},"/api/v1/auth/forget_password/{uuid}":{"get":{"tags":["AUTH"],"summary":"Get new password","description":"The user gets here by a link from the mail and receives an automatically generated access password.","operationId":"get_new_password_api_v1_auth_forget_password__uuid__get","parameters":[{"description":"Http link suffix for user registration","required":true,"schema":{"title":"Uuid","maxLength":32,"minLength":32,"type":"string","description":"Http link suffix for user registration"},"example":"993418d6434848bc927a13c48ca111f4","name":"uuid","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordScheme"}}}},"422":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}},"400":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}}}}},"/api/v1/auth/all":{"get":{"tags":["AUTH","ADMIN"],"summary":"Get all users","description":"Access for admin only","operationId":"get_all_auths_api_v1_auth_all_get","parameters":[{"required":false,"schema":{"title":"Offset","type":"integer","default":0},"name":"offset","in":"query"},{"required":false,"schema":{"title":"Limit","maximum":50.0,"type":"integer","default":10},"name":"limit","in":"query"},{"required":false,"schema":{"title":"Is Actve","type":"boolean"},"name":"is_actve","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get All Auths Api V1 Auth All Get","type":"array","items":{"$ref":"#/components/schemas/AuthModel"}}}}},"422":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}},"400":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/users/me":{"get":{"tags":["USERS","PROFILE"],"summary":"View a personal profile","operationId":"watch_me_api_v1_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseParentAuthScheme"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["USERS","PROFILE"],"summary":"Delete a personal profile","operationId":"delete_me_api_v1_users_me_delete","responses":{"204":{"description":"Successful Response returns only status code 204"}},"security":[{"OAuth2PasswordBearer":[]}]},"patch":{"tags":["USERS","PROFILE"],"summary":"Update a personal profile","operationId":"update_me_api_v1_users_me_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateParentScheme"}}},"required":true},"responses":{"200":{"description":"Successful Response returns only status code 200","content":{"application/json":{"schema":{}}}},"422":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}},"400":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/users/all":{"get":{"tags":["USERS","ADMIN"],"summary":"Get All Users","operationId":"get_all_users_api_v1_users_all_get","parameters":[{"required":false,"schema":{"title":"Offset","type":"integer","default":0},"name":"offset","in":"query"},{"required":false,"schema":{"title":"Limit","maximum":50.0,"type":"integer","default":10},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get All Users Api V1 Users All Get","type":"array","items":{"$ref":"#/components/schemas/ParentModel"}}}}},"422":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}},"400":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/users/{user_id}":{"get":{"tags":["USERS","ADMIN"],"summary":"Read User","operationId":"read_user_api_v1_users__user_id__get","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParentModel"}}}},"422":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}},"400":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/providers/me":{"get":{"tags":["PROVIDERS"],"summary":"Watch Me","operationId":"watch_me_api_v1_providers_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/providers/update":{"patch":{"tags":["PROVIDERS"],"summary":"Update Me","operationId":"update_me_api_v1_providers_update_patch","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/providers/delete":{"delete":{"tags":["PROVIDERS"],"summary":"Delete Me","operationId":"delete_me_api_v1_providers_delete_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/providers/":{"get":{"tags":["PROVIDERS"],"summary":"Get All Providers","operationId":"get_all_providers_api_v1_providers__get","parameters":[{"required":false,"schema":{"title":"Offset","type":"integer","default":0},"name":"offset","in":"query"},{"required":false,"schema":{"title":"Limit","maximum":50.0,"type":"integer","default":10},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}},"400":{"description":"Error","content":{"application/json":{"schema":{"type":"object","properties":{"detail":{"type":"string"}}}}}}}}},"/api/v1/providers/{user_id}":{"get":{"tags":["PROVIDERS"],"summary":"Read Provider","operationId":"read_provider_api_v1_providers__user_id__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Index","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"deprecated":true}}},"components":{"schemas":{"AuthModel":{"title":"AuthModel","required":["password","email"],"type":"object","properties":{"password":{"title":"Hashed password","maxLength":64,"type":"string"},"user_type":{"title":"User Type","type":"integer","description":"There are 3 params: user`1`, tutor`2`, institution`3`","default":1},"email":{"title":"Email","maxLength":254,"type":"string"},"id":{"title":"Id","type":"integer"},"is_active":{"title":"Is Active","type":"boolean","description":"Is the user activated","default":false}},"description":"Table for authentication data.\n\n#### Attrs:\n- id (int):\n    Iidentifier.\n- user_type (int):\n    UserType.\n- password (str):\n    Hashed pasword.\n- email (str):\n    User's email.\n- is_active (bool): Default False.\n    Is the user activated."},"Body_forget_password_api_v1_auth_forget_password_post":{"title":"Body_forget_password_api_v1_auth_forget_password_post","required":["email"],"type":"object","properties":{"email":{"title":"Email","maxLength":254,"type":"string"}}},"Body_login_api_v1_auth_token_post":{"title":"Body_login_api_v1_auth_token_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","maxLength":254,"type":"string"},"password":{"title":"Password","maxLength":32,"minLength":8,"type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"Body_request_change_password_api_v1_auth_change_password_post":{"title":"Body_request_change_password_api_v1_auth_change_password_post","required":["password"],"type":"object","properties":{"password":{"title":"Password","maxLength":32,"minLength":8,"type":"string"}}},"CreateTempUserScheme":{"title":"CreateTempUserScheme","required":["password","email"],"type":"object","properties":{"password":{"title":"Password for authorization","maxLength":32,"minLength":8,"type":"string","example":"qwe7RTY8asd"},"user_type":{"title":"User Type","type":"integer","description":"There are 3 params: user`1`, tutor`2`, institution`3`","default":1},"email":{"title":"User's email","type":"string","format":"email","example":"example@gmail.com"}},"description":"Scheme for first registration.\n\n#### Attrs:\n- email (str):\n    User's email.\n- user_type (int): Default `1`.\n    UserType.\n- password (str):\n    Password for authorization."},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ParentModel":{"title":"ParentModel","type":"object","properties":{"name":{"title":"Real user's name","maxLength":128,"type":"string","example":"John"},"surname":{"title":"Real user's last name","maxLength":128,"type":"string","example":"Doe"},"patronic":{"title":"Real user's patronic name","maxLength":128,"type":"string","example":"Ibn Abu Dakar"},"born":{"title":"User's date of birth","type":"string","format":"date"},"id":{"title":"Id","type":"integer"},"auth_id":{"title":"Auth Id","type":"integer"}},"description":"Table for parent data.\n\n#### Attrs:\n- id (int):\n    Identifier.\n- name: (str | None):\n    Real parent's name.\n- surname: (str | None):\n    Real parent's last name.\n- patronic: (str | None):\n    Real parent's patronic name.\n- born: (date | None):\n    Parent's date of birth as UNIX time.\n- auht_id (int | None):\n    Relation to athenticated data.\n    If None, the user is considered deleted."},"PasswordScheme":{"title":"PasswordScheme","required":["password"],"type":"object","properties":{"password":{"title":"Password for authorization","maxLength":32,"minLength":8,"type":"string","example":"qwe7RTY8asd"}},"description":"The base schema for creating password schemas.\n\n#### Attrs:\n- password (str):\n    Password for authorization."},"ResponseParentAuthScheme":{"title":"ResponseParentAuthScheme","required":["email"],"type":"object","properties":{"name":{"title":"Real user's name","maxLength":128,"type":"string","example":"John"},"surname":{"title":"Real user's last name","maxLength":128,"type":"string","example":"Doe"},"patronic":{"title":"Real user's patronic name","maxLength":128,"type":"string","example":"Ibn Abu Dakar"},"born":{"title":"User's date of birth","type":"string","format":"date"},"email":{"title":"Email","maxLength":254,"type":"string"}},"description":"Scheme for data of parent for issuing to the outside.\n\n#### Attrs:\n- name: (str | None):\n    Real parent's name.\n- surname: (str | None):\n    Real parent's last name.\n- patronic: (str | None):\n    Real parent's patronic name.\n- born: (date | None):\n    Parent's date of birth as UNIX time.\n- email (str):\n    Parent's email."},"TokenScheme":{"title":"TokenScheme","required":["access_token"],"type":"object","properties":{"access_token":{"title":"Access Token","type":"string","description":"JWT token"},"token_type":{"title":"Token Type","type":"string","default":"bearer"}},"description":"Scheme for `Bearer Token`.\n\n#### Attrs:\n- access_token (str):\n    JWT token.\n- token_type (str): Default `bearer`. Immutable."},"UpdateParentScheme":{"title":"UpdateParentScheme","type":"object","properties":{"name":{"title":"Real user's name","maxLength":128,"type":"string","example":"John"},"surname":{"title":"Real user's last name","maxLength":128,"type":"string","example":"Doe"},"patronic":{"title":"Real user's patronic name","maxLength":128,"type":"string","example":"Ibn Abu Dakar"},"born":{"title":"User's date of birth","type":"string","format":"date"}},"description":"Scheme for updating parent data."},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/token"}}}}}}